Create README.md
In this project we will work on visualizing the gender gap across college degrees. So far, we mostly focused on the STEM degrees but now we will generate line charts to compare across all degree categories.

We are going to look at the number of women in various majors between 1968-2011. We will then plot our findings to see if there are majors with less women than others.